Consolidate helper function from windows and libuv
